replace context with connect for notifications
